# openapi.yaml
openapi: 3.0.3
info:
  title: University API
  version: "1.0.0"
  description: >
    API for user authentication, course management, grade management, and user management.
servers:
  - url: http://localhost:5000/api/v1
    description: Local development server
paths:
...
  /users/:
    get:
      tags:
        - Users
      summary: List all users
      description: >
        **Requires JWT authentication.** Only Administrators can list all users.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of users returned successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
        "403":
          description: Unauthorized access.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "500":
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Users
      summary: Create a new user
      description: >
        **Requires JWT authentication.** Only Administrators can create new users.
      security:
        - bearerAuth: []
      requestBody:
        description: User creation payload.
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - password
                - role
              properties:
                name:
                  type: string
                  example: Jane Doe
                email:
                  type: string
                  format: email
                  example: jane.doe@example.com
                password:
                  type: string
                  example: "Password123!"
                role:
                  type: string
                  enum: [Student, Professor]
                  example: Student
      responses:
        "200":
          description: User created successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: User created successfully.
                  user:
                    $ref: '#/components/schemas/User'
        "400":
          description: Missing fields or invalid role.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: Unauthorized access.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "409":
          description: User already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "500":
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /users/{user_id}:
    get:
      tags:
        - Users
      summary: Get user details
      description: >
        **Requires JWT authentication.** Accessible by Admins, Professors, or the user themselves.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
          description: The unique identifier of the user.
      responses:
        "200":
          description: User details retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        "403":
          description: Unauthorized access.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - Users
      summary: Update a user
      description: >
        **Requires JWT authentication.** Accessible by the user themselves or an Administrator.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
          description: The unique identifier of the user.
      requestBody:
        description: Fields to update (name, email, password).
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: John Doe Updated
                email:
                  type: string
                  format: email
                  example: john.updated@example.com
                password:
                  type: string
                  example: "NewPassword123!"
      responses:
        "200":
          description: User updated successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: User updated successfully.
                  user:
                    $ref: '#/components/schemas/User'
        "403":
          description: Unauthorized access.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "500":
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - Users
      summary: Delete a user
      description: >
        **Requires JWT authentication.** Only Administrators can delete users.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
          description: The unique identifier of the user.
      responses:
        "200":
          description: User deleted successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: User deleted successfully.
        "403":
          description: Unauthorized access.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "500":
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: "user123"
        name:
          type: string
          example: John Doe
        email:
          type: string
          format: email
          example: john.doe@example.com
        role:
          type: string
          enum: [Administrator, Professor, Student]
          example: Student
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Error message"
      required:
        - error